thuthesis.dtx: use (xe)CJKfntef to underline bachelor title
To draw the underline for bachelor's thesis title, the old method
uses latex picture. For long title that spans two lines, the first
underline will be of constant length, which can make the output not
so good, especially when the title contains English words (as in the
sample), or when alternative manual line breaking is wanted.

Use \CJKunderline provided by (xe)CJKfntef can eliminate this issue.
The underline will be always along with the texts.

Note that xeCJKfntef will be called automatically when we require
CJKfntef using xelatex with ctex documentclass.